… __tests__/monitoring/performance-middleware.test.ts
… __tests__/monitoring/performance-monitor.test.ts
… components/dashboard/api-performance-monitor.tsx
… docs/API_PERFORMANCE_MONITORING.md
… docs/PERFORMANCE_MONITORING_IMPLEMENTATION.md
… lib/monitoring/performance-monitor.ts